+++ src/mediator/HandlerFn.ts
@@ -1,0 +1,2 @@
+export type HandlerFn = (...params: any[]) => void;
+export default HandlerFn;
