+++ src/mediator/index.ts
@@ -1,0 +1,3 @@
+import Mediator from './mediator';
+let mediator = new Mediator();
+export default mediator;
