+++ src/topics/IErrorWithDataId.ts
@@ -1,0 +1,5 @@
+interface IErrorWithDataId extends Error {
+  /** Identifier to use as a suffix on the 'error:' topic */
+  id?: string;
+}
+export default IErrorWithDataId;
